--- package/src/steps/ConfirmPaymentView/PaymentReview.models.ts
@@ -1,7 +1,0 @@
-import { NextStep } from "../../ApiContext";
-
-export type PaymentReviewProps = {
-  nextStep: NextStep & { type: "paymentReview" };
-  includeCryptoAddr?: boolean;
-  onButtonAction?: () => void;
-};
