+++ package/src/steps/PaymentReviewView/iconsMapping.tsx
